TST: split test_data to several files
