+++ pandas_datareader/tests/test_eurostat.py
@@ -71,7 +71,6 @@ class TestEurostat(tm.TestCase):
             result = df[expected.name]
             tm.assert_series_equal(result, expected)
 
-
 if __name__ == '__main__':
     nose.runmodule(argv=[__file__, '-vvs', '-x', '--pdb', '--pdb-failure'],
                    exit=False)
