+++ pandas_datareader/tests/test_oecd.py
@@ -79,7 +79,6 @@ class TestOECD(tm.TestCase):
         with tm.assertRaises(ValueError):
             web.DataReader(1234, 'oecd')
 
-
 if __name__ == '__main__':
     nose.runmodule(argv=[__file__, '-vvs', '-x', '--pdb', '--pdb-failure'],
                    exit=False)
