+++ pandas_datareader/tests/test_wb.py
@@ -234,7 +234,6 @@ class TestWB(tm.TestCase):
             self.assertTrue(result.columns.equals(exp_col))
             self.assertTrue(len(result) > 10000)
 
-
 if __name__ == '__main__':
     nose.runmodule(argv=[__file__, '-vvs', '-x', '--pdb', '--pdb-failure'],
                    exit=False)  # pragma: no cover
